% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Negate \%in\%}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{A vector}

\item{table}{A vector matching the mode of \code{x}}
}
\value{
A logical vector with length = \code{length(x)}
}
\description{
This function will return a logical vector that is the negation of \%in\%.
}
\examples{
c(1, 2, 3) \%nin\% c(3, 4, 5)

}
\keyword{internal}
