% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPO2-package.R
\docType{package}
\name{IPO2-package}
\alias{IPO2}
\alias{IPO2-package}
\title{IPO2: Parameter Optimization for XCMS Data Processing}
\description{
Isotopologue Parameter Optimization (IPO) was initially released as
    a package to facilitate the optimization of various XCMS parameters to 
    improve peak identification and retention time alignment. With the development
    of XCMS, the ease of integration with IPO diminished. IPO2 aims to apply a 
    similar approach to parameter optimization as the original IPO while working
    with the latest versions of XCMS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wmoldham/IPO2}
  \item Report bugs at \url{https://github.com/wmoldham/IPO2/issues}
}

}
\author{
\strong{Maintainer}: William Oldham \email{wm.oldham@gmail.com} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
