% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_centwave.R
\name{optimize_centwave}
\alias{optimize_centwave}
\title{Find centWave parameters}
\usage{
optimize_centwave(
  raw_data = NULL,
  parameter_list = suggest_centwave_params(),
  log_file = NULL,
  plot_dir = NULL
)
}
\arguments{
\item{raw_data}{MSnExp object containing the raw data}

\item{parameter_list}{List of parameters to set or to optimize}

\item{log_file}{Path for log file}

\item{plot_dir}{Path for plot directory}
}
\value{
A list of length 2 containing:
\item{history}{a data.table object recording the optimization history}
\item{best_cwp}{a CentWaveParam object of the optimal parameters}
}
\description{
\code{optimize_centwave} will identify parameters for \code{xcms::findChromPeaks} by
optimizing the identification of isotope pairs. Note that currently the
isotope identification algorithm relies on high resolution m/\emph{z} measurements.
}
\examples{


}
