% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_align_group.R
\name{optimize_align_group}
\alias{optimize_align_group}
\title{Find peak alignment and grouping parameters}
\usage{
optimize_align_group(
  xcmsnexp = NULL,
  parameter_list = suggest_align_group_params(),
  log_file = NULL,
  plot_dir = NULL
)
}
\arguments{
\item{xcmsnexp}{An XCMSnExp object}

\item{parameter_list}{List of parameters to set or to optimize}

\item{log_file}{Path for log file}

\item{plot_dir}{Path for plot directory}
}
\value{
A list of length 3 containing:
\describe{
\item{history}{a data.table object recording the optimization history}
\item{best_obi}{an ObiWarpParam object of the optimal parameters}
\item{best_density}{an PeakDensityParam object of the optimal parameters}
}
}
\description{
\code{optimize_align_group} will identify parameters for \code{xcms::adjustRtime} using
the ObiWarp algorithm and for \code{xcms::groupChromPeaks} using peak densities.
}
\examples{

}
