% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_centwave.R
\name{optimize_centwave}
\alias{optimize_centwave}
\title{Find centWave parameters}
\usage{
optimize_centwave(
  raw_data = NULL,
  parameter_list = suggest_centwave_params(),
  out_dir = NULL
)
}
\arguments{
\item{raw_data}{MSnExp object containing the raw data}

\item{parameter_list}{List of parameters to set or to optimize}

\item{out_dir}{Path for output directory. Defaults to working directory.}
}
\value{
A list of length 2 containing:
\describe{
\item{history}{a data.table object recording the optimization history}
\item{best_cwp}{a CentWaveParam object of the optimal parameters}
}
}
\description{
\code{optimize_centwave} will identify parameters for \code{xcms::findChromPeaks} by
optimizing the identification of isotope pairs. Note that currently the
isotope identification algorithm relies on high resolution m/\emph{z} measurements.
}
\examples{
\dontrun{
files <- list.files(system.file("extdata", package = "IPO2"), full.names = TRUE)
raw_data <-
  MSnbase::readMSData(files = files[[1]], mode = "onDisk") \%>\%
  MSnbase::filterRt(c(300, 900))
cwp_params <- optimize_centwave(raw_data = raw_data)
}

}
